+++ sample/src/commonMain/kotlin/com/github/panpf/zoomimage/sample/AppSettings.common.kt
@@ -90,6 +90,8 @@ expect class AppSettings(context: PlatformContext) {
 
     val subsamplingEnabled: SettingsStateFlow<Boolean>
 
+    val autoStopWithLifecycleEnabled: SettingsStateFlow<Boolean>
+
 
     // ------------------------------------------ other --------------------------------------------